+++ rasa_core/processor.py
@@ -16,7 +16,9 @@ from typing import Optional, List, Dict, Any
 from typing import Text
 
 from rasa_core.actions import Action
-from rasa_core.actions.action import ActionRestart, ACTION_LISTEN_NAME
+from rasa_core.actions.action import (
+    ActionRestart, ACTION_LISTEN_NAME,
+    ACTION_RESTART_NAME)
 from rasa_core.channels import UserMessage, InputChannel, channel
 from rasa_core.channels import CollectingOutputChannel
 from rasa_core.dispatcher import Dispatcher
@@ -40,7 +42,7 @@ try:
     scheduler = BackgroundScheduler()
     scheduler.start()
 except UnknownTimeZoneError:
-    logger.warn("apscheduler failed to start. "
+    logger.warning("apscheduler failed to start. "
                 "This is probably because your system timezone is not set"
                 "Set it with e.g. echo \"Europe/Berlin\" > /etc/timezone")
 
@@ -257,7 +259,7 @@ class MessageProcessor(object):
         if (num_predicted_actions == self.max_number_of_predictions and
                 should_predict_another_action):
             # circuit breaker was tripped
-            logger.warn(
+            logger.warning(
                     "Circuit breaker tripped. Stopped predicting "
                     "more actions for sender '{}'".format(tracker.sender_id))
             if self.on_circuit_break:
@@ -316,7 +318,7 @@ class MessageProcessor(object):
             if isinstance(e, SlotSet) and e.key not in slots_seen_during_train:
                 s = tracker.slots.get(e.key)
                 if s and s.has_features():
-                    logger.warn("Action '{0}' set a slot type '{1}' that "
+                    logger.warning("Action '{0}' set a slot type '{1}' that "
                                 "it never set during the training. This "
                                 "can throw of the prediction. Make sure to "
                                 "include training examples in your stories "
@@ -374,9 +376,9 @@ class MessageProcessor(object):
     def _save_tracker(self, tracker):
         self.tracker_store.save(tracker)
 
-    def _prob_array_for_action(self, action):
-        # type: (Action) -> Optional[List[float]]
-        idx = self.domain.index_for_action(action.name())
+    def _prob_array_for_action(self, action_name):
+        # type: (Text) -> Optional[List[float]]
+        idx = self.domain.index_for_action(action_name)
         if idx is not None:
             result = [0.0] * self.domain.num_actions
             result[idx] = 1.0
@@ -387,21 +389,21 @@ class MessageProcessor(object):
     def _get_next_action_probabilities(self, tracker):
         # type: (DialogueStateTracker) -> List[float]
 
-        follow_up_action = tracker.follow_up_action
-        if follow_up_action:
-            tracker.clear_follow_up_action()
-            result = self._prob_array_for_action(follow_up_action)
+        followup_action = tracker.followup_action
+        if followup_action:
+            tracker.clear_followup_action()
+            result = self._prob_array_for_action(followup_action)
             if result:
                 return result
             else:
                 logger.error(
                         "Trying to run unknown follow up action '{}'!"
                         "Instead of running that, we will ignore the action "
-                        "and predict the next action.".format(follow_up_action))
+                        "and predict the next action.".format(followup_action))
 
         if (tracker.latest_message.intent.get("name") ==
                 self.domain.restart_intent):
-            return self._prob_array_for_action(ActionRestart())
+            return self._prob_array_for_action(ACTION_RESTART_NAME)
 
         return self.policy_ensemble.probabilities_using_best_policy(
                 tracker, self.domain)