+++ rasa_core/remote.py
@@ -108,7 +108,7 @@ class RasaCoreClient(object):
         if response.status_code == 200:
             return response.json()
         else:
-            logger.warn("Got a bad response from rasa core :( Status: {} "
+            logger.warning("Got a bad response from rasa core :( Status: {} "
                         "Response: {}".format(response.status_code,
                                               response.text))
             return None
@@ -137,7 +137,7 @@ class RasaCoreClient(object):
                     logger.debug("Finished uploading")
                     return response.json()
             except RequestException as e:
-                logger.warn("Failed to send model upload request. "
+                logger.warning("Failed to send model upload request. "
                             "{}".format(e))
 
             if max_retries > 0:
@@ -146,7 +146,7 @@ class RasaCoreClient(object):
                 time.sleep(2)
 
         if response:
-            logger.warn("Got a bad response from rasa core while uploading "
+            logger.warning("Got a bad response from rasa core while uploading "
                         "the model (Status: {} "
                         "Response: {})".format(response.status_code,
                                                response.text))