+++ rasa_core/trackers.py
@@ -13,6 +13,7 @@ from typing import Generator, Dict, Text, Any, Optional, Iterator
 from typing import List
 
 from rasa_core import events
+from rasa_core.actions.action import ACTION_LISTEN_NAME
 from rasa_core.conversation import Dialogue
 from rasa_core.events import (
     UserUttered, ActionExecuted,
@@ -69,7 +70,7 @@ class DialogueStateTracker(object):
         # if tracker is paused, no actions should be taken
         self._paused = None
         # A deterministically scheduled action to be executed next
-        self.follow_up_action = None
+        self.followup_action = ACTION_LISTEN_NAME
         self.latest_action_name = None
         self.latest_message = None
         # Stores the most recent message sent by the user
@@ -103,6 +104,7 @@ class DialogueStateTracker(object):
             "slots": self.current_slot_values(),
             "latest_message": self.latest_message.parse_data,
             "latest_event_time": latest_event_time,
+            "followup_action": self.follow_up_action,
             "paused": self.is_paused(),
             "events": evts
         }
@@ -314,7 +316,7 @@ class DialogueStateTracker(object):
         self.latest_action_name = None
         self.latest_message = UserUttered.empty()
         self.latest_bot_utterance = BotUttered.empty()
-        self.follow_up_action = None
+        self.follow_up_action = ACTION_LISTEN_NAME
 
     def _reset_slots(self):
         # type: () -> None
@@ -357,11 +359,11 @@ class DialogueStateTracker(object):
 
         self.follow_up_action = action
 
-    def clear_follow_up_action(self):
+    def clear_followup_action(self):
         # type: () -> None
         """Clears follow up action when it was executed"""
 
-        self.follow_up_action = None
+        self.followup_action = None
 
     def _merge_slots(self, entities=None):
         # type: (Optional[List[Dict[Text, Any]]]) -> List[SlotSet]